h265: The code stream with width and height of 64 failed the CTS test. [1/1]
PD#SWPL-17038

Problem:
When the width and height of the decoded macro block is 1,
the HEVC_PARSER_LCU_START parameter read from the register is 0,
which will cause the decoded data to be set to error.

Solution:
(hevc->lcu_x_num * hevc->lcu_y_num != 1)
Add the above judgment conditions to unset the decoded data to Error.

Verify:
P212

Change-Id: Ia89f5381059ef6d996bb8c60408324e0509e5b2d
Signed-off-by: Gan Zhang <gan.zhang@amlogic.com>