media_module: h264 Update the ref queue correctly [1/1]
PD#OTT-7782

Problem:
When there is some error data in the stream, the error ref frame is
forced to be released. Since the list of ref frames is not updated,
decoder driver thinks that the DPB buffer is incorrect and the entire
DPB queue to be cleared, result frame drop.

Solution:
This problem is resolved by updating the list of refs when the
incorrect ref frames are removed to dpb queue.

Verify:
u212

Change-Id: I754434ae1daaaf8df56e9efc10233ca7725df86e
Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>