media_module: Fixed h265 playback crash [1/1]
PD#SWPL-23625

Problem:
On the GXL, ucode loaded by 4k and 1080p in
h265 format is different, but we think
that the same ucode is loaded by h265,
so the different ucode is not loaded repeatedly,
which leads system restart.

Solution:
This problem is solved by reloading
the ucode each time when the decoder
instance switch.

Verify:
u212

Change-Id: I5f4c105a19e72199eee7276813bcad6a6e07a1bd
Signed-off-by: Yixin Peng <yixin.peng@amlogic.com>