media_module: fixed h264 playback stuck issue [1/1]
PD#SWPL-24328

Problem:
If I frame is followed by frame B of the
previous GOP in the bitstream, it cause
that the B frame cannot find enough
reference frames after the operation of
buffer reset, resulting in reset the
buffer repeatedly, causing playback stuck.

Solution:
For this type of B frame, since the B
frame will not be displayed, we no need
to detect the frame reference relationship,
so the reset buffer will not be operation
repeatedly.

Verify:
U212

Change-Id: I311ea0787a4258b71f8f38b20247c4303c51a36c
Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>