vmh264: fix video jitter bug [1/1]
PD#SWPL-24276

Problem:
Interlaced stream jitter.

Solution:
This type of stream consists of a pack of two fields
and a pack of single fields.
When the second field is the same as the pts of the first field,
the display module will drop the second field, causing jitter.
The pts of the second field needs to be set to 0,
and the display module reconstructs the pts
through duration / 2 to output the second field.

Verify:
u212

Change-Id: Ice9e802d9d24183392a47aa982eab8be79f0b1ad
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>