vmh264: fix bug of stream not fluent [1/1]
PD#SWPL-26687

Problem:
Multi-instance h264 stream not fluent.

Solution:
The code stream has two sps but the id is 0.
As a result, the sps syntax is not updated,
and the original frame_mbs_only_flag is used,
resulting in an error.
update original frame_mbs_only_flag when reset.

Verify:
u212

Change-Id: I3d731b73d504d4bd9403382e088fcb31c64fd15f
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>