vmh264: fixed the MB count error on decoding. [1/1]
PD#SWPL-27454

Problem:
the source of h264 with ippp playback abnormal.

Solution:
has some wrong with first I frame so will be checked
on the decoding, then the frame will be show if the
error rate below 5%, but now the mb with calculation is
incorrect which cause it do not a chance to display.

Verify:
w200

Change-Id: I865dd6bec6a8aa9bd830f9800afdf037d7843e23
Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>