h264: stream mode vpts lookup fail [1/1]
PD#SWPL-35520

Problem:
Because the stream will checkin one vpts every three frames, and some
of the frames size is small, it will cause lookup pts is successful
but it doesn't match, resulting playback unsmooth.

Solution:
When lookup fail is detected many times and the frame size is
small, pts is computed using duration to solve the problem.

Verify:
SC2

Change-Id: Ib9319fc46e28006ccca173c87faf88b88288ba66
Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>