v4l: fix green stripes at the bottom of the video. [1/1]
PD#SWPL-39682

Problem:
As the size value configured for HEVC_SAO_Y_LENGTH and HEVC_SAO_C_LENGTH
registers is smaller than the value of relevant parameters in canvas
configuration, the decoded output data only fills in part of the data,
and the subsequent data does not fill in the Canvas buffer.

Solution:
Align the scaled width and height to 64 bytes, then multiply to get size.
Configure the size to HEVC_SAO_Y_LENGTH and HEVC_SAO_C_LENGTH registers
to ensure that the size will not be smaller than the size configured in canvas.

Verify:
AH212

Change-Id: Ibd0eb369e6724b9509e7f6fbe1bab12ba201b30f
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>