media_module: h265 the first frame is different when play many times [1/1]
PD#SWPL-39208

Problem:
CSD information is discarded mistakenly due to the dirty data found at
the start of playing, cause this problem.

Solution:
When it is found that when the dirty data is more than 1M at start
play, it will not discarded completely, and only half of the data is
discarded to solve this problem.

Verify:
SC2

Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>
Change-Id: Ia498f5b97496a7f392fcd5547d088a5224ff6eb0