media_module: fixed h265 crash issue about aux data release [1/1]
PD#SWPL-40912

Problem:
Due to this video file contains suffix of sei, the aux buffer will
reallocation at decode next frame, but the release of the aux buffer
was used on the back-end before the reallocation, causing crash issue.

Solution:
This problem is solved by sending the frame to display after all
the suffix aux data are ready.

Verify:
SC2

Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>
Change-Id: Ibb3f4d4dc01985e50b527a628aa142c200478c05