v4l: fix av1 freeze when buffer is  not enough. [1/1]
PD#SWPL-42540

Problem:
When there are more than one frame in the stream,
the run ready condition is changed, causing the
decoding buffer to get stuck.

Solution:
When there are multiple frames in the stream,
you need to wait buffer form the omx2.
There is no need to change the number of buffers for run ready.

Verify:
AH212

Change-Id: I83d5050a58e2d59c719a46165713f14c90c6e3f0
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>