v4l: optimize fb buffer allocation. [1/1]
PD#SWPL-43303

Problem:
optimize the fb buffer allocation with token.

Solution:
fixed the issue that dec and vpp alloction fb buffer at the
same time. the previous was that dec got the token of fb,
then vpp was in mutex state, which would cause to cost more
time when start playback. now this issue was fixed, dec & vpp
allocate are independet of each other, so as long as there
are idle fb in m2m pool, it can be allocated.

Verify:
ah212

Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>
Change-Id: I72e0cca591ae137d5626468548e2c5a68b7a4da1