dv: The screen will flash black during the resolution switching. [1/1]
PD#SWPL-44910

Problem:
The decoding of the DV module fails when the data stored in the
DV is zeroed during the video resolution switching process.

Solution:
The variables in the ctx structure are used to save the dv data,
so that the dv data will not be released during the video resolution
switching process.

Verify:
ohm

Change-Id: Ib60a55eefc2a9f9413145ee494353472be90699a
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>