h265: fix video is freeze. [1/1]
PD#SWPL-44742

Problem:
As the decoded data is detected to be over decoder, it is forced to set
to error.As a result, the following frame is not output after the frame
is also set to error.

Solution:
When decoder mode is frame mode, over decoder is not detected.

Verify:
Ohm

Change-Id: I7b69d08cad1176870339df24b51f0d0cd80ac49e
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>