av1: CB1 fix some dv streams playback flash screen. [1/1]
PD#SWPL-42874

Problem:
The dv data is not cleared, resulting in data error.

Solution:
When there are multiple frames in a chunk, the previous dv data
needs to be cleared. It is necessary to clear the DV data of the
available buffer when there is a buffer to continue decoding.

Verify:
ohm

Change-Id: I8d092b4b0cd0c55ec6a56245df48268ffaebadfb
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>