mpeg12: fix pts and crc fail in cts. [1/1]
PD#SWPL-45626

Problem:
Because the decoded data after the VPP module can not guarantee
that the data out each time is the same.

Solution:
When a buffer is allocated in the upper layer of OMX2, the upper
layer passes a flag to the driver. The driver uses this flag to
output one of the two fields without passing through the VPP.

Verify:
ohm

Change-Id: I9f42f1aa431d843d03997f0c5683e3b8b28d1148
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>