v4l: fix some fail in cts. [1/1]
PD#SWPL-45049

Problem:
1.android.security.cts.StagefrightTest#testStagefright_cve_2017_0728
There is a mistake in the above case stream. Two buffers are needed
to decode in the process of starting the broadcast. When there is still
a decoding buffer, the dpb will not be flushed causing it to get stuck.

2.android.security.cts.StagefrightTest#testStagefright_cve_2017_0765
The resolution switch occurred because the width of the stream exceeded
1920 and was forced to 1920.

Solution:
1.Change run_ready_min_buf_num to one in the broadcast start process,
and force flush dpb when certain conditions are met.

2.The width and height information is not modified.

3.An error is reported to the upper when the width or height of the
stream is less than 16.

Verify:
ohm

Change-Id: I6ae188392a3b6f588455bef1a9171a118b83c90d
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>