av1: fix some dv streams playback flash screen. [1/1]
PD#SWPL-41229

Problem:
The dv data is not cleared, resulting in data error.

Solution:
When there are multiple frames in a chunk, the previous dv data
needs to be cleared. It is necessary to clear the DV data of the
available buffer when there is a buffer to continue decoding.

Verify:
ohm

Change-Id: I45a6dd054f26d6bdbbcca87a7f140761361d8452
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>