h265: fix dv streams playback flash green screen. [1/1]
PD#SWPL-45977

Problem:
There is no PORT_TYPE_DUALDEC flag in V4L2, after parsing NAL_EOS,
the dv data is not stored in the buffer of the current frame.

Solution:
When NAL_EOS is parsed in the dv streams, the dv data should be stored
in the buffer of the current frame.

Verify:
ohm

Change-Id: I4840b4f5686b997efd9091676eb79b3a4ff0352b
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>