h264: fix dv stream struck. [1/1]
PD#SWPL-45972

Problem:
Due to some special dv streams of h264, the decoded data will not be
output until it is larger than the reorder, which will cause a
probability stuck.
The logic of dv stream is different from that of normal outflow frame.
There is a small probability of frame loss.

Solution:
1.If it is dv stream, need to add an additional buffer.
2.The dv stream does not enter the logic line of error correction
processing.

Verify:
ohm

Change-Id: I4c5dd884060c6a335f1363ae5b825be708df262b
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>