media_module: CF1 fixed decoder fence issue [1/1]
PD#SWPL-48631

Problem:
1.the user speace drop frame caused decoder
fence to be returned early, the decoder
did not handle it accordingly, cause playback
stuck.
2.decoder fence related memory has been
freed but is is still in use, cause kernel crash.

Solution:
1.For the frames returne early, it will be
stored in a queue of fence, and the franme
will not be put to newqueue until the fence
is signaled.
2.Fixed this problem by adding a notification
mechanism release the fence related memory
when the back-end releases all buffers.

Verify:
SC2

Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>
Change-Id: I95b21c039a69e2a24595603aa2c06aaa394d9744