hdr10p: the serial port prints stack information. [1/1]
PD#SWPL-48786

Problem:
The presence of vfree in the function of vh265_vf_put may cause sleep.

Solution:
When the hdr10p data size is less than 128, memory is allocated
to the hdr10p data through kzmalloc. And finally freed through kfree.
Ps:Since the upper layer can only pass a maximum of 200 bytes of
parameters, 128 bytes are used to pass hdr10p data to the upper layer,
which is only for CTS authentication.

Verify:
T3

Change-Id: Ic7e57d20194be17ed88b73cbcd6dbcd2a00d1b33
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>