v4l2: fixed crash due to abnormal exit. [1/1]
PD#SWPL-51522

Problem:
If the video gets stuck, restart the system through reboot,
which will cause the fd of the di buffer to be closed,
and the driver will destroy the pipeline first. Since
the vpp module has not been destory, the di module will
send the released di buffer to the vpp module for output.
There was a crash due to pipeline being destroyed.

Solution:
When an abnormal exit occurs, the vpp must be destroyed before
the pipeline can be destroyed.

Verify:
T5D

Change-Id: I957e5bc2cf184db762e0ef5eb406b72aa6e7e8c1
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>