v4l2: fix vpp buffer size error in special case. [1/1]
PD#SWPL-50493

Problem:
When reorder_margin is changed to 2 on the first start, seek immediately
causes the buffer size to be recalculated. Since reorder_margin is
modified, this will cause the vpp buffer size to be
calculated incorrectly.

Solution:
dec margin and vpp margin are calculated and processed.

Verify:
am311

Change-Id: I71ca273c28231f2054f6714ed473be2c12967454
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>