vp9: CB1 fixed issue of reports resolution change. [1/1]
PD#SWPL-52535

Problem:
Youtube vp9 crash.
Two consecutive headers were sent when the resolution was changed,
resulting in repeated reporting of the resolution change message,
and at the same time, the buffer was referenced
without obtaining the buffer id,
resulting in a crash.

Solution:
Advance the time point of updating the width and height
of the last resolution to
after the resolution is changed, instead of updating the width and height
of the last resolution after outputting the first frame,
and set the status to init when the resolution is changed.

Verify:
AH212

Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>
Change-Id: I06cd9f27d99a203a6fbbe7071d7c683f23b4f7ba