AV1: CB1 Seek freeze when streams has errors [1/1]
PD#SWPL-53182

Problem:
When there are multiple frames in one packet,
and there are errors in the stream packet,
the watch dog will not work and cause a stuck.
The reason is that jiffies are not recorded
when decoding a packet of multiple frames.

Solution:
Set the jiffies when start decoding multiple frames in one packet.

Verify:
ah212

Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>
Change-Id: I8234ad0ed96b6220681372845c45b851586c7cf4