h264: CB1 fixed playback stuck after seek. [1/1]
PD#SWPL-51966

Problem:
Waiting for the buffer in a packet of multiple frames, if the seek
operation is performed, the decoding driver enters the again state,
so that the start of the seek after the seek is abnormal due to the
fact that the decoder is not stopped.

Solution:
When entering the agin state, you need to stop the decoder.

Verify:
T5D

Change-Id: I4a0a6a7f4f8b644f412d81deb8f69ced0081c87f
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>