media_module: CB1 h264 fixed playback stuck when resolution change. [1/1]
PD#SWPL-58546

Problem:
After resolution change, if the total number of
buffers belong to the decoder (before resolution
change buffer + after  resolution change buffer)
exceeds a certain number, decoding will not to
decode.The problem is that when such a situation
is found, the decoder does not remove some unused
buffers from the DPB, causing the decoder thinks
that there are no usable buffers.

Solution:
If it is found that total number of buffers
belong to the decoder exceeds the limit, it
will check whether any buffers can be removed
to solve the problem.

Verify:
U212

Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>
Change-Id: I0c8b09c7d1aa6b698bde0fc5aaffcddf54120450