v4l2: CB0 vp9 mosiac when seeking. [1/1]
PD#SWPL-57739

Problem:
When decoding the current vp9 frame,
it depends on the previous frame and the comv of the previous frame.
When abnormal, the frame buffer of the previous frame is released,
resulting in a blurry screen.

Solution:
Ensure that two adjacent frames do not use the same frame buffer.

Verify:
ah212

Change-Id: I70fba53e08eabb28233383b2b07f260f2b45006b
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>