vp9: CB1 Vp9 frame mode decoder error [1/1]
PD#SWPL-64858

Problem:
When the decoder cannot solve the frame data,
it will obtain the frame data of the next frame,
and usually the frame with show_existing_frame flag
only contains one frame header and does not contain
the frame data. When multiple frames with show_existing_frame
flag appear consecutively, the decoder will be stuck without
obtaining the frame data for a long time.

Solution:
When decoded to a frame that has the show_existing_frame flag,
the decoding simply ends without attempting to parse the frame data

Verify:
ohm

Signed-off-by: kuan.hu <kuan.hu@amlogic.com>
Change-Id: I2f22b9a933f0644444736ed48a230771799a6ea6