av1: CB1 fixed playback stuck after reset. [1/1]
PD#SWPL-66750

Problem:
Since the vdec_core_finish_run function is called twice, the status will
eventually be set to connect. During the reset connect process, it will
return if the status is not equal to disconnect, and the playback will
be stuck if there is no decoding instance.

Solution:
The timer function does not need to schedule work after detecting the
decoding state to prevent two work from appearing in one run.

Verify:
ohm

Signed-off-by: kuan.hu <kuan.hu@amlogic.com>
Change-Id: I06f04b6e118cc29f3cda12027ad21e865758d660