decoder: CB1 fix the playback fail caused by stream errors in codec2. [1/1]
PD#SWPL-69169

Problem:
There was an error in the stream, which caused the decoder to output no
frames, resulting in a playback failure.

Solution:
In code2, the upper layer will set the corresponding flag to the
decoding driver. After the flag is recognized in the decoding, it will
be marked with the error flag and output data to the upper layer.

Verify:
ohm

Change-Id: I24dd02f7774bd90a58cb353b614ecaa9e00eb182
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>