av1: CF2 afbc buf for lesss memory usage in fg mmu scene. [1/1]
PD#SWPL-77888

Problem:
Save scatter memory in fg mmu scene (dw = 0x21).

Solution:
In the fg scene, there are two types of afbc buffers,
one is for fg, the other is not.
afbc buffers (not for fg) do not participate
in the cycle of the entire video channel,
but only rotate inside the decoder;
so the afbc buffers (not for fg) can be released early
when dereferenced, thereby saving scatter memory.

Verify:
s4

Change-Id: Ic6bfd161591612a96267ac44dea62fbe8060b607
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>