mpeg12: CB2 fix output pts error in AndroidS CTS Test. [1/1]
PD#SWPL-81925

Problem:
The pts information of the decoded B frame is 0,
The backend will calculate the pts of the B frame
according to the duration, so that the calculated
pts information will be incorrect and cause the test
to fail.

Solution:
The decoded output pts information uses the one provided by
the upper layer.

Verify:
Franklin

Signed-off-by: kuan.hu <kuan.hu@amlogic.com>
Change-Id: I7ace38ebe14a7cd217c08c0fbadcc27872083526