decoder: CB2 fix dvb playback crash. [1/1]
PD#SWPL-83212

Problem:
For the interlace stream, v4lvideo uses the index of
the vf output by di to get the aux data, and the aux data
corresponding to this index has been reused by the decoder,
and free is called during the decoding and reuse process,
so there is a certain probability of a crash.

Solution:
For the interlace stream, it is not allowed to call the
decoder callback function to get aux data.

Verify:
ohm

Change-Id: I82784d1d5eadbc5b1df8051d43eec3ed15f6d812
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>