h265: CB2 DV stream playback unsmooth in stream mode. [1/1]
PD#SWPL-83597

Problem:
Because the pts of the stream is abnormal, the pts_unstable set by the
upper layer was cleared by the second decoder, resulting in a jump in
the pts output in the decoding here.

Solution:
Copy the sys_info parameter passed by the upper layer to the second
decoder, and then use the pts_unstable passed by the upper layer to
correct the transition of pts.

Verify:
SC2

Change-Id: Ia241d5fdbfcac6c67f89dab28f271ece34cb05c0
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>