H265/AV1: CB2 freeze when playback h265/av1. [1/1]
PD#SWPL-75270
PD#SWPL-88231

Problem:
1.h265:The bar code stream syntax value
sps_max_dec_pic_buffering_minus1_0 is 15,
the theoretical dpb size is 18,
and the current driver limits the dpb to 16,
so the buffer is insufficient and stuck.
2.av1:After enabling lowlatency,
the frame is output in the current decoding round
and will be placed in the outframe array.
In the next round of decoding,
the ref_count of the outframe will be decremented by 1 between
put_mv_buf and get_free_fb. When the round is used,
it will cause the mv buffer to leak.

Solution:
1.Set dpb cap max_buf_num to 18.
2.After the ref_count of outframe is reduced by 1,
put_mv_buf is performed again
before get_free_fb to avoid mv buffer leakage.

Verify:
sc2

Change-Id: Ida6bbcf78e0a5d9f7aa6277ec8cfbfc5b4056479
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>