v4l: CB2 fix video playback stuck and mosaic. [1/1]
PD#SWPL-84886

Problem:
1.The vpp module does not get the buffer, causing the playback to
be stuck.
2.When switching the resolution, since the number of buffers is used to
configure the pic according to the number of buffers calculated
according to the stream, but in fact the upper layer will send some more
buffers down, because there is no configuration and the decoding obtains
a pic without configuration, the playback display is abnormal.

Solution:
1.When decoding the output eos, let the vpp module get the buffer.
2.When switching resolutions, use the number of buffers configured by
the upper layer to reconfigure the pic information.

Verify:
ohm

Change-Id: I008629c9d3cc64a65a6cb6981ec3bc6e3fbcf4b8
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>