v4l: CB2 fix playback abnormal problem. [1/1]
PD#SWPL-101845

Problem:
Since the upper layer sends the buffer first and then the start
streaming, this will cause the vdec to not schedule and cause
the playback to be stuck.

Solution:
When the state of ctx is AML_STATE_ACTIVE, call the up function to start
the vdec thread to start the decoder decoding.

Verify:
ohm

Change-Id: I632365ef7c4f66af7c9a7898cf52f9f019451b9b
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>