vdec: CB2 fix playback abnormal problem. [1/1]
PD#SWPL-102185

Problem:
At the beginning of playback, the input data is not enough,
causing the vdec to not schedule and get stuck.

Solution:
When the input data is not enough, set the corresponding
bit of need_more_data to 1 to let the upper layer send
data down and schedule vdec to decoder.

Verify:
S5

Change-Id: I114c8a03d3cab3f3d2346557ca0de9d5c237a3ca
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>