h265: CB1 fix decoding exception due to stream error. [1/1]
PD#SWPL-103442

Problem:
When the decoding is nal done, first call the fb_hw_status_clear
function, and then read the mmu0_ptr and mmu1_ptr in the
read_bufstate_front function to be 0, so that the hardware will
write the 0 address in the next decoding, causing the memory to crash.

Solution:
When the decoding is nal done, first call the read_bufstate_front
function to read mmu0_ptr and mmu1_ptr, and then call the
fb_hw_status_clear function.

Verify:
S5

Change-Id: I5118eef03b7f7e8cfa45daa2f1b9ed51c1557ea2
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>