v4l: CB1 dw mmu box dynamic allocation. [1/1]
PD#SWPL-104615

Problem:
Since the dw mmu box is allocated at the start of the broadcast,
when the scene where the fg stream is switched to a non-fg stream,
it will not be able to play due to unallocated memory.

Solution:
Add dw mmu box dynamic allocation function.

Verify:
s4/s4d

Change-Id: I633d3d0a0c5c61431965ec3df39f8533e2806218
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>