decoder: CB2 fix the problem of multi-channel playback freeze. [1/1]
PD#SWPL-106142

Problem:
Avs modified the AV_SCRATCH_K register to other values during the
decoding process, when h264 of the vdec core are decoded,
it will be thought that the driver has opened the debug information
of the ucode. In this way, interrupts will be triggered continuously,
causing the playback to freeze.

Solution:
Configure the AV_SCRATCH_K register every time.

Verify:
S5

Change-Id: I7d7a22c063547c89a815c190f8087ce96e156a35
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>