avs2: CB2 The avs2 probability decoding exception problem. [1/1]
PD#SWPL-101713

Problem:
The avs2 probability parses to the header position of the next frame,
which will cause decoding exceptions in the current frame.

Solution:
Remove the logic of not checking buffer empty in dual core mode. When
parsing to the end of this frame of data, compare the parsed size
information with the size information of the frame, so as to prevent
parsing from being overdone.

Verify:
S5

SourceCode:
ucode:
0.4.25-gc50c02f
I0eae0
Ic13c4
I265cc
I7fef4
I30648

Change-Id: I0eae012b0793ce45e57537d0cad7da4bc9ac3021
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>