decoder: CB2 fix the playback exception caused by hardware exceptions. [1/1]
PD#SWPL-109349

Problem:
Decoding the wrong stream leads to hardware problems. In the next
run, a forced reset will cause decoding exceptions.

Solution:
When decoding run, first disable the front or back arbiter, and then
reset the action, so as to ensure that there are no executed commands in
the arbiter.

Verify:
S5

Change-Id: I222087f994c1105be64793d6fbf8d8a41d29fa60
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>