vp9: CB2 fix playback problems caused by scheduling exceptions. [1/1]
PD#SWPL-109724

Problem:
Since the decoding releases the hardware first, and then calls the
interface for allocating memory, this will cause the times of isr and
isr_fn to be inconsistent, so that the vdec has been scheduled and the
system is stuck.

Solution:
First call the interface that allocates memory, and then release
the hardware to ensure that it will not be stuck in isr_fn.

Verify:
S5

Change-Id: Id5f75a43c4e35ca60e8720e2312b9ef6469aeacf
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>