avs3: CB2 fix playback stuck problem due to packet loss. [1/1]
PD#SWPL-115612

Problem:
When packet loss occurs on the network, the decoding will lack the
reference frame.In this case, the reference relationship of the previous
frame will be updated, resulting in an abnormal backref.

Solution:
When an exception occurs, there is no need to update the reference
relationship of pic.

Verify:
S5

Change-Id: I97d4d279b822ed760d3e14b0356ff91cf9638e5b
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>