avs3: CB2 fix playback stuck problem due to decoding exception. [1/1]
PD#SWPL-115666

Problem:
Since the sys_imem_ptr address of the front is updated when decoding the
header,when the same frame of data is decoded multiple times, the
sys_imem_ptr will be abnormal. Therefore, after back gets the wrong
instruction, it will decode the exception.

Solution:
It is not necessary to update the sys_imem_ptr address of the front when
the header is decoded, it needs to be updated after the decoding of the
front pic is completed.

Verify:
S5

Change-Id: Ice8892950f73807b5a463526a340f0b3d70e9b00
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>