v4ldec: CB2 Capture MMAP mode crash. [1/1]
PD#SWPL-137449

Problem:
The di post version modifies the hash table key,
replaces the physical address with the dma buffer,
and the dma buffer in capture mmap mode is empty,
causing an exception.

Solution:
The mmap mode still uses the physical address as the key,
and judges whether the dma buffer is empty.

Verify:
ohm

Change-Id: I69a8f8e0925910dbd2e2ace0d08bed490441819e
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>