h264: CB2 fix playback aliasing problems for interlace stream. [1/1]
PD#SWPL-146145

Problem:
For the intrelace stream, the type_original variable does not set the
corresponding interlace information to the aisr module, which causes the
aisr module to not process the intrelace source and cause playback
aliasing.

Solution:
Correctly set the source scan information to the type_original variable.

Verify:
S5

Change-Id: I19adf4aabd6094ed346a136b4e70480c9d9a6e75
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>