vp9: CB2 Correction for new buffer management [1/1]
PD#SWPL-149998
PD#SWPL-147129

Problem:
1.The vp9/av1 error stream causes buffer reference confusion
and abnormal operations.
2.When the interlaced stream is abnormal,
a frame with a pic struct of 0 appears,
causing the frame to be both output and error-recovered,
resulting in a ref exception and a crash.

Solution:
1.Perform buffer recycling through error branches
and add list validity detection to avoid crashes.
2.Only the interlaced frames of pair pic
enter the recycling logic of interlace_clear_no_pair_pic.

Verify:
S905X4

Change-Id: I82d9fc47a15587d9f169c835bafd7325c562975d
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>