h265: CB2 Blurry screen in surface texture mode [1/1]
PD#SWPL-150002

Problem:
h265/vp9/av1:
The avbc of the surface texture scene will be recycled in advance,
but the frame has been sent to the video instead,
causing the avbc to be displayed,
thus causing the screen to be distorted.

Solution:
Remove the avbc flag through the decoder
to avoid being used for video display.

Verify:
S905X4

Change-Id: I49f24b2f2fe3be67c9ca1b9e4f8ec269716c3d74
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>