h265: CB2 CtsMediaBitstreamsTestCases h265 case fail [1/1]
PD#SWPL-150744

Problem:
The bitstream is a frame sequence with i p i,
first_ Pic_ Flag was not assigned a value of 2
when decoding the second frame,
resulting in the second frame not being flushed out.

Solution:
The flag needs to be assigned a value of 2
after decoding the second frame.

Verify:
S905X4

Change-Id: I56971a3fdac9534f6be8ae4d83fe50c50589a575
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>