264: CB2 fix Some exceptions occur after the special flow seeks. [1/1]
PD#SWPL-161577

Problem:
When playing a 264 stream and performing a seek,
if the SPS data is not re-sent after the seek action,
it will cause the playback to fail because the SPS cannot
be found after the seek.

Solution:
If the sps information is not passed after seek,
the sps information before seek will be used to
ensure normal playback.

Verify:
T5M

Change-Id: Ie3bc565a3b897ec5d6246043c6eeb181199b35a1
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>