h265: CB2 fix stuck probability in H265 v4l stream mode. [1/1]
PD#SWPL-183505

Problem:
When using V4L2 streaming mode to play H265 DV source,
pic done does not set decoding_pic to NULL.
If again appears in the decoding of the next frame at this time,
the decoding information of decoding_pic will be cleared,
so the buff of the frame cannot be released normally.

Solution:
Set decoding_pic to NULL when pic is done.

Verify:
ohm

Change-Id: Iabf722c60913281c113dd9e600865ccd7d548670
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>