h264: CB2 fix scheduling information exceptions after turning on mmu mode. [1/1]
PD#SWPL-187647

Problem:
When mmu is turned on in h264,  the vdec_combine_flag variable in the
scheduling information  cannot be effectively cleared to 0 when decoding
exits.  As a result, subsequent playback cannot schedule vdec and hevc
decoding at the same time.

Solution:
When the hardware information used for decoding changes, call
vdec_core_release first and then vdec_core_request to ensure that
the vdec_combine_flag variable in the scheduling information is normal.

Verify:
Ohm and T6D

Change-Id: I7ef5887078b5e4124c250a59a1c3eb203b7a3505
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>