vdec: CB2 fix fence not signaled when error stream and resetting [1/1]
PD#SWPL-191802

Problem:
In fence_mode,
the frame header is decoded and then the buffer is posted
to the backend, but decoder resets when decoding frame body
because of some error in stream, so the fence is not signaled,
caused display wait fence timeout.

Solution:
when the buffer is posted to the backend, but fence is not signaled,
triggering fence signal once in work.

Verify:
S905Y5

Change-Id: I983ac25341d4fd6f8bbd11f654319b3b120df96a
Signed-off-by: xiaohang.cui <xiaohang.cui@amlogic.com>