vdec: CF2 Add error-frame dropping in fence mode [1/1]
PD#SWPL-177199

Problem:
in fence mode, the frame buffer is posted to display when
frame header is decoded. But only the frame body is decoded, we can know
the frame is error or not. So the error frame can not be dropped and it
will show mosaic.

Solution:
in fence mode, drop error frame in display.
Display will keep the showing frame and stop keeping when next
showenable frame comed. It needs decoder to check the fence statuts and
refcount to get the buffer status.

Verify:
SC2

Change-Id: I5f8ad9e65320bf4d0812e21618fb90370d0c283e
Signed-off-by: xiaohang.cui <xiaohang.cui@amlogic.com>