av1: CB2 Fixed the issue of the AV1 watchdog not starting in time. [1/1]
PD#SWPL-204771

Problem:
Since the AV1 driver’s watchdog was not started when searching for
the frame header after seq head done, it could not be triggered
when search_head encountered a corrupted stream, causing the
process to hang for an extended period.

Solution:
Enable the watchdog after seq head done.

Verify:
S6

Change-Id: Ia6bc03beb9612fe05443484739aada9f42647c09
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>